Initial Commit!
Note:
This was meant to explain that I am writing this after the date says, and that the commits were added after their date, in 2021-2022. I didn't track the old changes from 2018-2020 in Git (at all, to be honest :( ), so I am adding them later to make sure as much of the history can be kept around. I didn't keep backups of the pack as development of it went along, so I added as much of the updates as I could find online, through Internet Archive, and MediaFire. If you like rambly stories, this next block of text should be great for you! I started writing about the history of Dark Mode, and more about why these changes are getting added at a later date. Cheers!

So, crazy story ahead!

I started working on, what would become Dark Mode, in around mid-to-late 2017. I had just started looking into resource packs at that time, and I had no idea what I was doing! I was just getting a feel for how to change things in-game, like I had always wanted to do.

A little time later, I had started working on my own resource pack, Reality Pack. This was my first actual pack, and I haven't made it public at all (yet? It's not very good, haha). It had the goal of updating all of the textures in-game to have a high-resolution and realistic look to them. This was my project for most of the summer during 2017.

From around late 2017 to mid-2018, I had started working on Faithful 32x for Bedrock Edition! This was what really helped me with learning how to make resource packs. Seeing how other people made their packs really paved the way for my own projects in the future!

I made a Google Site for my Faithful projects (which I later used for all of my projects in general), from which I hosted my pack source on Google Drive from. This made it super easy to share the pack source, as I just had to upload a new `.mcpack` to the Google Drive folder for the update to be released. We'll get back to this later though... haha

Here's an Internet Archive link of the site I luckily randomly decided to make back then, which came in handy looking back on it:
https://web.archive.org/web/2018*/https://sites.google.com/view/faithful-32x-bedrock-edition/

While still working on Faithful 32x, I next moved on to Dark Mode and Simple Sides, around the summer of 2018. I worked a ton to get most of the features started for Dark Mode, and I got pretty much everything supported just over the summer. A crazy project to get done! I was so happy with it.

After working on those to get them more polished, I toyed with the idea of releasing them to the public, something that seemed daunting at the time. I wasn't new to the Internet or anything, but I hadn't really made a presence online before, so these (my resource packs) were my first big thing that I made into an online project! I'm so happy that I committed to moving forward with them.

After posting them to my own Google Site, I went to MCPEDL, which is where I had been learning about resource packs from thus far. First I posted Faithful 32x, then Dark Mode, and lastly Simple Sides.
https://web.archive.org/web/20181015000000*/https://mcpedl.com/faithful-32x-pack/
https://web.archive.org/web/20180815000000*/https://mcpedl.com/dark-mode-resource-pack/
https://web.archive.org/web/20180915000000*/https://mcpedl.com/simple-sides

So, after those notes to make note of, there's the point that I started using Google Drive to share my pack files. Not really a bad thing to do that, but it was worse that I didn't keep any of the updates along the way as I would make changes. So, yeah.

This is me, coming up on about nearly 4 years (4, WHAT??? wow) since making the pack, scrambling to find old versions of my resource pack throughout the web, so I can add them back into my repo that I started about two years after that. Who'da thunk it? That's a funny scenario! The lesson: DON'T DELETE YOUR BACKUPS, IT'S NOT WORTH IT.

So yeah, lost it a little. This commit is authored on August 6, 2018. Well, I'm actually writing this on April 18, 2022, which is when this was actually added to the repo. Really messy, I know. For me, I think it will be more important to have all of the history of the source code, rather than a chunk of it missing because it hadn't been tracked yet.

The first real commit to the repo will be the commit "Initial Upload", as that was when I first made the repo on GitHub, "back" in 2020 (back, being reference to when I wrote this, 2022).

Okay so the jist of all this, please don't delete your backups while you can have them. I have been scrambling through Internet Archive saves and MediaFire links to find any remnants of Dark Mode from the pre-Git days, and it has been confusing and difficult. Before my first version uploaded to MCPEDL, I have no history to show for that. It's gone to the wind. The only echoes that are left of it are what's still in the pack today. It's not that big of a deal, as I still have everything that came after that, but I don't have any of the early days.

It's kind of confusing that a ton of the textures will show up in the next commit, that's not how it went.

I am so glad that I learned from this though. Had I not kept all of my versions of Smart Text Editor around from the early days, that would've been a bummer. I think I'm reflecting on this so much because I'm looking at the old versions of the project before I learned all of my own programming techniques, and it's painful! I will have to do the same process for Simple Sides, and I surely can't wait to do that, haha! It would be great to have all of the features saved from over the years, it was just in reach.

Welp, this should be a blog or something. Sorry for the long commit message. Maybe you got something out of it?

If you read all of that, great job. Make sure to save your code history to your repository before brushing your teeth, it'll keep the cavities away!

Signing off! (You should listen to "Somewhere In This Universe Someone Hits A Drum", I found that back in 2020, and it is a FANTASTIC album!)
Brandon Bennett