Bedrock Files Commit
Note: This commit and message were actually added to the repo at a later date, on April 18, 2022. Questions? Visit the mess of a story about that in the root commit just one step back XD

Alrightey, so this is the first commit with the source of the project.

At this point I had just recently published the project to MCPEDL, so this appears to be the first publicly-available release of Dark Mode!

The version was working great with the fairly recent v1.5.0 release for Bedrock at the time. It was Phase Two of the Update Aquatic! I have to say, that was one of my favorite times of working on resource packs.

It also seems I had (recently? not sure to be honest) added support for Faithful 32x! This was a huge step for the pack, as it supported both Vanilla textures, and for Faithful! I later changed this to only support Vanilla, but maybe I will look into it again in the future, now that my dev tooling has gotten much better to handle it (Git, GitHub, Photopea, all of the modern things of today).

It also didn't have support for Java yet, that released about a year and a half later during 2020. It took a while for me to develop it, and then I finally finished it up and release it during the extra time that came with the pandemic.

Cool! Much covered there too. I don't plan for the changelogs to be too log for every single one of these. Just thought I'd help explain some of the gaps in history for the missing timeline ahead.