The Faithful Fiasco
Alr', 'ere it is. 'Ere ain't 'neugh room in 'is town fer the two urv us...

So, luckily as of writing this (still in 2022, read the first commit in the repo for background of why that is), AntVenom greatly covers why Faithful is removed from Dark Mode at this point. Props to him for making this video! I can't believe I have a role in it too :O
https://www.youtube.com/watch?v=kgZYBTlwaVs

Now I am reminded of why I stopped working on Faithful after this point in the timeline. I am glad that I moved on to other things after that! My part in Faithful also re-converges back in with the EM32-Bedrock (Emulated Vattic => Classic Faithful => Faithful again :) ) community project in late 2021, where the end of the xMrVizzy era looms! As of this message on April 19, 2022, the Compliance and Classic Faithful teams have merged and emerged together under the original Faithful name! I'm so glad that everyone is finally together again, and that the correct ownsership of the Faithful brand came back. It really was a shame that it went away for some time. Can't wait to work with the modern Faithful team again in future updates!
https://faithfulpack.net/

This is where I contributed by Faithful project from this point in the timeline over to the modern Faithful team (Emulated Vattic / Classic Faithful, back in late 2021)
https://github.com/ClassicFaithful/32x-Jappa-Bedrock