Added System Files to .gitignore
Kinda weird, I know. This was originally a normal commit that added the `.gitignore` when the date says. That's all normal. Well, it was re-added back into the past with the first commit of the repo with a rebase to add historical versions of the pack. Check out "Initial Commit!" to learn more about this wacky file that was added, which already exists. I thought it would be just as important to show that this was when the file was really added to the repo.