The Monumental Rebase
Alright, can't believe I just pulled that off. I rebased all of Dark Mode's pre-Git history to the start of the repo. With correct dates, changelogs, and everything. That took a long time to do. I had to work on it every once in a while over the last few months.

If you want to learn about why it was so hard to do, go to the first commit, "Initial Commit!". I wrote a way-too-long few lines about the Dark Mode journey, and how it sums up to today!

Sorry for the lack of content updates recently, this had been the main reason why. I have really been wanting the get the rest of the project's history into the repo, one way or another.

Some of the updates along the way, through mid-2019, were lost to the fire, but I think I caught a great deal of the amount of time and effort that came before the repo too :)

Thanks for coming along for this wild ride, seeya again soon! I really gotta start going to bed earlier.
Brandon Bennett

Oh yeah, and the changelog for this one:

Changes:
- Updated the README to the current year! Welcome to 2022, in April XD