Furnace Screen Updates
This commit formats a few parts of the Furnace Screen, and it adds a factory entry for locked item notifications, which it looks like I removed with my original edit to the file. Not sure where/when it's used exactly, but it's probably better to keep it there rather than to just remove it.