Min Engine Version
Moving back the Min Engine Version metadata since it doesn't need to be all the way up to the bleeding edge. I'm running Bedrock with the Unix Launcher version, and that is currently only on version 1.20.1, so it wasn't letting Dark Mode load properly, because the Min Engine Version was set higher than that, at 1.20.30. This still seems to be a realistic bump up from 1.13.0 too though, since that's already about 4 years ago now. If you're going back to old versions like that (you probably aren't), then using the specific Dark Mode release for that version is the better way to go anyways. I usually only want to upgrade the Min Engine Version when the pack adds non-compatible features with new releases, or it removes features no longer needed in newer releases. I think that may happen here pretty soon with 1.20 changes, so that's why I'm upgrading it this time around.

There are a ton more changes between this and Java, a LOT has changed in terms of the UI stack (moreso for Java). I think I just need to ensure that Bedrock's current code reflects that of the vanilla game's declarations still, then it'll be back to good again. Java needs a complete full-rehash unfortunately. I didn't even update it since after the Nether Update yet either, there's a bunch it could use some help with.